fix: update code lens position when resolving
This is only a semi-fix as the server will still send back the old code
lens position until the entire file is elaborated and the resulting
ilean info sent to the watchdog again. Even if the worker would send
incremental ilean updates, this would not fix the problem.

In theory, the response to the code lens resolution request would need
to be delayed until the worker reaches the item's position in the source
file, or if the respective definition was removed, until the worker
reaches the end of the file.